Toggle IC/OOC chat & save content, not just replace to 'OOC '
