+++ sources/client/representation/qt/gaminglineedit.h
@@ -10,8 +10,11 @@ class GamingLineEdit: public QLineEdit
 public:
     GamingLineEdit(QWidget* parent = 0);
     void keyPressEvent(QKeyEvent* event);
+    void ToggleOOC();
 signals:
     void keyToPass(QKeyEvent* event);
 private:
     bool is_arrow_pass_;
 };
+
+bool IsOOCMessage(const QString& text);
